[UHYU-433] refactor: brand, guest 도메인 주석 제거 및 코드 리팩토링
