+++ src/main/java/com/ureca/uhyu/domain/admin/dto/response/StatisticsMembershipUsageRes.java
@@ -1,5 +1,6 @@
 package com.ureca.uhyu.domain.admin.dto.response;
 
+import com.ureca.uhyu.domain.admin.dto.MembershipUsageByBrand;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
