+++ src/main/java/com/ureca/uhyu/domain/brand/dto/request/UpdateBrandReq.java
@@ -3,7 +3,6 @@ package com.ureca.uhyu.domain.brand.dto.request;
 import com.ureca.uhyu.domain.brand.dto.BenefitDto;
 import com.ureca.uhyu.domain.brand.enums.StoreType;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 import java.util.Optional;
