+++ src/main/java/com/ureca/uhyu/domain/brand/dto/response/CategoryListRes.java
@@ -1,7 +1,9 @@
 package com.ureca.uhyu.domain.brand.dto.response;
 
 import com.ureca.uhyu.domain.brand.entity.Category;
+import io.swagger.v3.oas.annotations.media.Schema;
 
+@Schema(description = "카테고리 목록 조회 응답 dto")
 public record CategoryListRes (
         Long categoryId,
         String categoryName
