+++ src/main/java/com/ureca/uhyu/domain/brand/dto/response/CreateUpdateBrandRes.java
@@ -1,7 +1,9 @@
 package com.ureca.uhyu.domain.brand.dto.response;
 
 import com.ureca.uhyu.domain.brand.entity.Brand;
+import io.swagger.v3.oas.annotations.media.Schema;
 
+@Schema(description = "브랜드 생성, 수정 응답 dto")
 public record CreateUpdateBrandRes(
         Long brandId
 ) {
