+++ src/main/java/com/ureca/uhyu/domain/brand/dto/response/InterestBrandRes.java
@@ -1,7 +1,9 @@
 package com.ureca.uhyu.domain.brand.dto.response;
 
 import com.ureca.uhyu.domain.brand.entity.Brand;
+import io.swagger.v3.oas.annotations.media.Schema;
 
+@Schema(description = "관심 브랜드 목록 조회 응답 dto")
 public record InterestBrandRes (
         Long brandId,
         String brandName,
