+++ src/main/java/com/ureca/uhyu/domain/brand/repository/BrandRepository.java
@@ -5,7 +5,6 @@ import com.ureca.uhyu.domain.brand.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface BrandRepository extends JpaRepository<Brand, Long>
         , BrandRepositoryCustom{
