+++ src/main/java/com/ureca/uhyu/domain/guest/dto/response/GuestRecommendationRes.java
@@ -1,5 +1,8 @@
 package com.ureca.uhyu.domain.guest.dto.response;
 
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "비로그인 사용자 추천 응답 dto")
 public record GuestRecommendationRes (
         Long brandId,
         String brandName,
