+++ src/main/java/com/ureca/uhyu/domain/user/repository/history/HistoryRepositoryCustomImpl.java
@@ -3,7 +3,7 @@ package com.ureca.uhyu.domain.user.repository.history;
 import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.ureca.uhyu.domain.admin.dto.response.StatisticsMembershipUsageRes;
-import com.ureca.uhyu.domain.admin.dto.response.MembershipUsageByBrand;
+import com.ureca.uhyu.domain.admin.dto.MembershipUsageByBrand;
 import com.ureca.uhyu.domain.brand.entity.QBrand;
 import com.ureca.uhyu.domain.brand.entity.QCategory;
 import com.ureca.uhyu.domain.store.entity.QStore;
